Extract Scripts into its own module
